refactor: BaseOptimizer
