+++ pytorch_optimizer/optimizer/lookahead.py
@@ -4,7 +4,7 @@ from typing import Dict
 import torch
 from torch.optim import Optimizer
 
-from pytorch_optimizer.base.base_optimizer import BaseOptimizer
+from pytorch_optimizer.base.optimizer import BaseOptimizer
 from pytorch_optimizer.base.types import CLOSURE, DEFAULTS, LOSS, OPTIMIZER, STATE
